chore(youtube): rename endpoint files to kebab-case
